+++ src/renderer/renderer.hpp
@@ -39,15 +39,15 @@ class RenderAPI {
 
     void flush();
 
-    void draw(const Triangle &triangle, const int layer = 0);
+    void draw(const Triangle &triangle, const float layer = 0.0f);
 
-    void draw(const Point &line, const int layer = 0);
+    void draw(const Point &line, const float layer = 0.0f);
 
-    void draw(const Line &line, const int layer = 0);
+    void draw(const Line &line, const float layer = 0.0f);
 
-    void draw(const Quad &quad, const int layer = 0);
+    void draw(const Quad &quad, const float layer = 0.0f);
 
-    void draw_text(const std::string &text, const float x, const float y, const float height, const int layer = 0);
+    void draw_text(const std::string &text, const float x, const float y, const float height, const float layer = 0.0f);
 
     void clear();
